ajustando testes

+++ app/Http/Controllers/TypeUserController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\AbilitiesEnum;
+use App\Models\TypeUser;
 use App\Repositories\TypeUserRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use App\Http\Requests\TypeUserRequest;
 
@@ -12,35 +15,60 @@ class TypeUserController extends Controller
     {
     }
 
-    public function store(TypeUserRequest $request):JsonResponse
+    /**
+     * @throws AuthorizationException
+     */
+    public function store(TypeUserRequest $request): JsonResponse
     {
-        $payload= $request->validated();
-        $typeUser=$this->typeUsersRepository->create($payload);
+        $this->authorize(AbilitiesEnum::VIEW, TypeUser::class);
+
+        $payload = $request->validated();
+        $typeUser = $this->typeUsersRepository->create($payload);
         return response()->json($typeUser, 201);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function show(string $id): JsonResponse
     {
-        $typeUser=$this->typeUsersRepository->findById($id);
+        $this->authorize(AbilitiesEnum::VIEW, TypeUser::class);
+
+        $typeUser = $this->typeUsersRepository->findById($id);
         return response()->json($typeUser, 200);
     }
 
-    public function update(string $id, TypeUserRequest $request):JsonResponse
+    /**
+     * @throws AuthorizationException
+     */
+    public function update(string $id, TypeUserRequest $request): JsonResponse
     {
-        $payload= $request->validated();
-        $typeUser=$this->typeUsersRepository->update($id, $payload);
+        $this->authorize(AbilitiesEnum::VIEW, TypeUser::class);
+
+        $payload = $request->validated();
+        $typeUser = $this->typeUsersRepository->update($id, $payload);
         return response()->json($typeUser, 201);
     }
 
-    public function destroy(string $id):JsonResponse
+    /**
+     * @throws AuthorizationException
+     */
+    public function destroy(string $id): JsonResponse
     {
+        $this->authorize(AbilitiesEnum::VIEW, TypeUser::class);
+
         $this->typeUsersRepository->delete($id);
         return response()->json([], 204);
     }
 
-    public function index():JsonResponse
+    /**
+     * @throws AuthorizationException
+     */
+    public function index(): JsonResponse
     {
-        $typeUser=$this->typeUsersRepository->listAll();
-        return response()->json($typeUser, 200);
+        $this->authorize(AbilitiesEnum::VIEW, TypeUser::class);
+
+        $typeUser = $this->typeUsersRepository->listAll();
+        return response()->json($typeUser);
     }
 }
